move switch + stories to input folder
